pyproject.toml
[project]
name = "baymesh"
version = "0.1.0"
description = "Baymesh command line and library for members of Bayme.sh"
readme = "README.md"
requires-python = ">=3.11"
maintainers = [
    {name = "Greg Taylor"},
]
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1.8",
    "meshtastic>=2.5.9",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "pyserial>=3.5",
    "requests>=2.32.3",
]

[project.scripts]
baymesh = "baymesh.cli:cli"

[project.urls]
Repository = "https://github.com/gtaylor/baymesh-cli"

[tool.uv]
package = true

[tool.ruff]
line-length = 90

[dependency-groups]
dev = [
    "pyright[nodejs]>=1.1.391",
    "pytest>=8.3.4",
    "python-semantic-release>=9.15.2",
    "ruff>=0.8.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
include = ["src", "tests"]
reportMissingTypeStubs = false

[tool.semantic_release]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"
version_variables = ["src/baymesh/__init__.py:__version__"]

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

